+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+* allow changing event title/desc in first 24h ([#65](https://github.com/liveask/liveask/issues/65))
+
 ## [2.8.0] - 2024-02-25
 
 ### Changed